Renamed  'stashvar_cmip6.py' to 'stashmasters.py'.
